requirements.txt
appdirs==1.4.4
audioread==3.0.0
azure-ai-textanalytics==5.2.1
azure-cognitiveservices-speech==1.24.1
azure-common==1.1.28
azure-core==1.26.0
azure-storage-blob==12.14.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
decorator==5.1.1
ffmpeg==1.4
greenlet==2.0.0
idna==3.4
isodate==0.6.1
joblib==1.2.0
librosa==0.9.2
llvmlite==0.39.1
msrest==0.7.1
numba==0.56.3
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
pooch==1.6.0
pycparser==2.21
pydub==0.25.1
pyodbc==4.0.34
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
soundfile==0.11.0
SQLAlchemy==1.4.42
threadpoolctl==3.1.0
typing-extensions==4.4.0
urllib3==1.26.12